[F, #2] Move content
Avoiding long files.
MC LogUploader -> Uploader Extensions
